break reciter screen into separate components
